All hail the first successful build under Python 3 (CPython3.6)

+++ Florence/Solver/FEMSolver.py
@@ -574,7 +574,7 @@ class FEMSolver(object):
                 # function_space, mesh, material, self, Eulerx, Eulerp)
 
             ParallelTuple = parmap.map(formulation,np.arange(0,nelem,dtype=np.int32),
-                function_space, mesh, material, self, Eulerx, Eulerp, processes=multiprocessing.cpu_count() / 2)
+                function_space, mesh, material, self, Eulerx, Eulerp, processes= int(multiprocessing.cpu_count()/2))
 
         for elem in range(nelem):
